requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
av=10.0.0=pypi_0
black=21.4b2=pypi_0
blas=1.0=mkl
braceexpand=0.1.7=pypi_0
brotli=1.0.9=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.9.24=ha878542_0
cachetools=5.3.1=pypi_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
contourpy=1.0.5=pypi_0
cudatoolkit=11.1.1=ha002fc5_10
cycler=0.11.0=pypi_0
dataclasses=0.6=pypi_0
datasets=2.8.0=pypi_0
decord=0.6.0=pypi_0
detectron2=0.6+cu111=pypi_0
dill=0.3.6=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
easydict=1.10=pypi_0
einops=0.5.0=pypi_0
exifread-nocycle=3.0.1=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.8.0=pypi_0
fire=0.4.0=pypi_0
focal-frequency-loss=0.3.0=pypi_0
fonttools=4.37.4=pypi_0
freetype=2.10.4=h0708190_1
freia=0.2=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=pypi_0
future=0.18.3=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.23.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.58.0=pypi_0
huggingface-hub=0.10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=pypi_0
imageio=2.22.2=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
img2dataset=1.41.0=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.4=pypi_0
lame=3.100=h7f98852_1001
langdetect=1.0.9=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuv=1.43.0=h7f98852_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.4.4=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
mpi4py=3.1.4=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
mutagen=1.46.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
networkx=2.8.7=pypi_0
ninja=1.11.0=h924138e_0
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
oauthlib=3.2.2=pypi_0
olefile=0.46=pyh9f0ad1d_1
omegaconf=2.3.0=pypi_0
opencv-python=4.6.0.66=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1o=h166bdaf_0
opentsne=0.6.2=pypi_0
packaging=21.3=pypi_0
pandas=1.5.0=pypi_0
pathspec=0.11.2=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.2.0=pypi_0
pip=22.2.2=py38h06a4308_0
portalocker=2.7.0=pypi_0
promise=2.3=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
pyarrow=7.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pypi_0
pycryptodomex=3.18.0=pypi_0
pydot=1.4.2=pypi_0
pyheif=0.7.1=pypi_0
pynvml=11.4.1=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.9.0=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2022.4=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.9.13=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rsa=4.9=pypi_0
scenedetect=0.6.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.1=pypi_0
seaborn=0.12.0=pypi_0
sentry-sdk=1.25.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=59.5.0=pypi_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0=pypi_0
smmap=5.0.0=pypi_0
soundfile=0.11.0=pypi_0
sqlite=3.39.3=h5082296_0
tabulate=0.9.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
timeout-decorator=0.5.0=pypi_0
timm=0.6.11=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pypi_0
torch-cluster=1.5.9=pypi_0
torch-geometric=2.2.0=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.12=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu111
tqdm=4.64.1=pypi_0
transformers=4.22.2=pypi_0
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.12=pypi_0
video2dataset=1.1.0=pypi_0
wandb=0.12.21=pypi_0
webdataset=0.2.48=pypi_0
websockets=11.0.3=pypi_0
webvtt-py=0.4.6=pypi_0
werkzeug=2.3.7=pypi_0
whatimage=0.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=3.2.0=pypi_0
xz=5.2.6=h5eee18b_0
yacs=0.1.8=pypi_0
yarl=1.8.2=pypi_0
yt-dlp=2023.3.4=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.12=h5eee18b_3
zstd=1.4.9=ha95c52a_0
